autorun: add SPDK_RUN_FUNCTIONAL_TEST=1
This runs some additional functional tests from the autotest test
suite. This is also a prerequisite to enabling NVMe tests.

For now, just leave the jobs named with "unittest". We may want to
rename this at some point. I think it will be awhile before we have
a heterogenous set of runners, where different runners are needed to
run specific parts of autotest. So for now I think it's best to just
leave everything as one job, enabling all of the tests we want to run
in the conf file for that one job.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ie51c6d9881f3a15d0e974d62347c668dac07888e