gerrit_changes_to_github: don't skip patches that already have branches
It is possible we might have branches in the CI mirror repo that
don't have a vote yet on Gerrit. For example, we have 30-40 of
these currently because they were used for early testing before
we started voting. It could also happen in the future if we run
a patch through autorun but then the vote POST fails due to Gerrit
being down.

So we still need to do autorun on these patches, we just need to
skip the branch creation part.

Signed-off-by: Jim Harris <jim.harris@samsung.com>