fix(gerrit_changes_to_github): replace 'in' with 'endswith'
When checking whether a branch already existed, then it did so by
checking whether all existing branch names were "in" the candidate
branch-name.

  existing("changes/95/24495/1") in "refs/changes/95/24495/1"

This is fine until patchsets exceed 9, then one would get True for:

  existing("changes/95/24495/1") in "refs/changes/95/24495/13"

Thus, to fix this, then this is replace with "endswidth".
For example:

  "refs/changes/95/24495/13".endwith("refs/changes/95/24495/1")

Thus, it will fail where it would previously return true due to the
subset match.

Signed-off-by: Simon A. F. Lund <os@safl.dk>